Iteration vs transition intervan
Hey!

I realized that the iteration interval for the loop, as well as the transition interval were both being set by interval, that being the case, it means that the interval does act as you would expect IE: you are not able to to control how long it sits on any one value, since the loop time and transition time are the same, its constantly changing.

This seems to be a good solution to allow more control cleanly.